pyproject.toml
[tool.poetry]
name = "politylink-tools"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
politylink = "^0.1.28"
requests = "^2.24.0"
tqdm = "^4.50.0"
# needed to add scrapy for politylink-crawler because poetry env switching did not work with subprocess
scrapy = "2.2.1"
boto3 = "^1.16.47"
google-cloud-speech = "^2.0.1"
pydub = "^0.24.1"
jinja2 = "^2.11.2"
opencv-python = "^4.5.1"

[tool.poetry.dev-dependencies]
jupyterlab = "^2.2.8"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
